independent_6/requirements.txt
google-cloud-storage
google-cloud-speech
ffmpeg-python
PyPDF2
